fix: 유저 삭제 수정

+++ src/main/java/com/vinny/project/user/UserService.java
@@ -86,6 +86,9 @@ public class UserService {
     }
 
     public void delete(String id){
+        if(!userRepository.existsById(id)){
+            throw new BusinessException(ErrorCode.USER_NOT_FOUND);
+        }
         userRepository.delete(id);
     }
 }
